Added changes requested by Alasdair in regards to the camera zooming behavior. Camera target object can now be changed. There are variables to play around with to get desired effect such as speed and angle restrictions. I also moved the script to appropriate place.
