+++ Empire - The Last Battle/Assets/Unity/Scripts/MonoBehaviours/CameraMovement.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
 guid: 5fd92a72694459146903be2f569008d0
-timeCreated: 1449334523
+timeCreated: 1449413289
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 0
+  executionOrder: 100
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
